Add launch screen for Android
- Modularize structure of multiple pages
